Install Flutter manually, not using a github action
Both actions I've found don't support Linux on arm64. The reason is flutter builds on
linux-arm64 are not official. So, we switch to installing Flutter manually.

Also, we remove the hardcoded flutter version from pubspec.yaml, as the only reason we put
it there is for the old github action to extract it from there. The new one doesn't even have such
a functionality.