Refuse to create / modify a prefix or pin an executable while apps are running
The reason is that the complications associated with running multiple apps at the same that are described in README.md may lead to WineBar handing when doing the operations in question.

For modifying a prefix or pinning an executable, we do allow apps running in other prefixes, except when wine is run under muvm (that is on Apple silicon).

For creating a prefix, the restrictions apply only when running on Apple silicon.