Stop using the -e argument of muvm to pass env vars to child process
Muvm doesn't propagate the environment variables to the process it starts in the virtual machine. Therefore, we used to pass the environment variabes to the child process by giving those environment variables as arguments to muvm (the -e NAME=VAL syntax). However, it turned out the environment variables specified that way in one muvm invocation may be visible in another [1].

Therefore, now we pass those environment variables as arguments to log-capturing-runner rather than
to muvm.

[1]: https://github.com/AsahiLinux/muvm/issues/206