Install the Flutter SDK through snap
Installing Flutter from snap means we can't control the version we are installing, but on the
positive side, the version from snap comes with older glib / gtk / whatever, so the binaries
it builds are able to run on older distros.

Unfortunately, the Flutter from snap brings an old version of CMake with it. We build a few C/C++
utilities in addition to our Flutter app and we need CMake 3.28 at least. To overcome this issue,
we download and extract a newer version of CMake and overlay it on top of the Flutter installation
from snap.

Yet another complication coming from the snap version of Flutter is that it sets various environment
variables such as LDFLAGS that break our win32-apps build. We solve that by explicitly unsetting those
variables in ExternalProject_Add().